FW: Add ignore-test-failure mode.
In this mode the framework is expected to ignore errors raised by the tests.
Although no error is reported if any tests fails, the failures are counted in ignored_fails counter for each thread.
When the test run is concluded a log with information can be printed to the console to infrom which threads failed and how many times.

Log example:

```
ok   n executed_test
 ---
  info: {version: <version>, timestamp: <timestamp>}
  Thread N on CPU M (pkg i, core j, thr k, family/model/stepping xx-xx-xx, microcode 0xnnnnnnn, PPIN N/A):
   - 'I> Found ignored failures. Number of ignored failures nnnn'
```

Signed-off-by: dawid-sabat <dawid.sabat@intel.com>